Rework of parenting
Parenting has been entirely changed and the way that positions are set have also been reworked.

Now when a child object is added, its transform is parented properly. It will also be updated when the parent's transform is changed.

Text Component now has its own Texture Component that does the rendering for it. No more duplication of code.